+++ CHANGELOG.md
@@ -11,6 +11,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Unit tests for awsbatch.py.
 
+### Removed
+
+- Test action for Python Version 3.10.
+
 ## [0.4.0] - 2022-07-28
 
 ### Added